moved AgenciesActionTypes to types, added loader, changed selector, added marging button to router-wrapper
