--- src/components/AgenciesPage/actions/ActionTypesAgencies.ts
@@ -1,5 +1,0 @@
-export const enum AgenciesActionTypes {
-  FETCH_AGENCIES_REQUEST = 'FETCH_AGENCIES_REQUEST',
-  FETCH_AGENCIES_SUCCESS = 'FETCH_AGENCIES_SUCCESS',
-  FETCH_AGENCIES_FAILURE = 'FETCH_AGENCIES_FAILURE',
-}
