+++ src/components/AgenciesPage/selectors/selectors.ts
@@ -1,9 +1,7 @@
 import { useSelector } from 'react-redux';
-import { Agency } from '../types/Agencies';
+import { AgenciesState } from '../types/Agencies';
 import type { StoreState } from '../../../store';
 
-export const selectAgencies = (): { agencies: Agency[] } => {
-  return useSelector<StoreState, { agencies: Agency[] }>(
-    state => state.agencies,
-  );
+export const selectAgencies = (): AgenciesState => {
+  return useSelector<StoreState, AgenciesState>(state => state.agencies);
 };
