+++ src/components/AgenciesPage/types/Agencies.ts
@@ -1,5 +1,8 @@
-import type { AgenciesActionTypes } from '../actions/ActionTypesAgencies';
-
+export const enum AgenciesActionTypes {
+  FETCH_AGENCIES_REQUEST = 'FETCH_AGENCIES_REQUEST',
+  FETCH_AGENCIES_SUCCESS = 'FETCH_AGENCIES_SUCCESS',
+  FETCH_AGENCIES_FAILURE = 'FETCH_AGENCIES_FAILURE',
+}
 export interface Agency {
   id: number;
   name: string;
