+++ src/components/Router/Router.scss
@@ -1,5 +1,5 @@
 .router-elements-wrapper {
-  margin: 0 auto;
+  margin: 0 auto 4rem;
   padding-top: 6rem;
   max-width: 96rem;
 }
